Bumping up dependencies
