+++ main/actions/src/main/scala/sbt/Package.scala
@@ -8,7 +8,7 @@ import java.io.File
 import java.util.jar.{ Attributes, Manifest }
 import collection.JavaConversions._
 import sbt.internal.util.Types.:+:
-import sbt.io.Path._
+import sbt.io.syntax._
 import sbt.io.IO
 
 import sbinary.{ DefaultProtocol, Format }
