+++ sbt/src/sbt-test/project/auto-plugins/project/Q.scala
@@ -1,8 +1,6 @@
 package sbttest // you need package http://stackoverflow.com/questions/9822008/
 
-	import sbt._
-	import Import._
-	import sbt.Keys.{name, resolvedScoped, organization	}
+	import sbt._, syntax._, Keys._
 	import java.util.concurrent.atomic.{AtomicInteger => AInt}
 
 object Imports
