+++ sbt/src/sbt-test/project/binary-plugin/changes/define/A.scala
@@ -1,8 +1,6 @@
 package sbttest // you need package http://stackoverflow.com/questions/9822008/
 
-import sbt._
-import Keys._
-import Import._
+import sbt._, syntax._, Keys._
 
 object C extends AutoPlugin {
 	object autoImport {
