+++ sbt/src/sbt-test/project/session-update-from-cmd/project/Common.scala
@@ -1,6 +1,4 @@
-import sbt._
-import Keys._
-import Import._
+import sbt._, syntax._, Keys._
 
 object Common {
   lazy val k1 = taskKey[Unit]("")
